Working correctly for the limited scale (same as in-class3, but with the option to simulate instead of running the physical hardware).
pt-ach script now automatically launches all the processes based
on whether or not the physical or virtual platform (default
virtual) is selected.

The only thing not done on the virtual platform is making the
ball move. The plugin for that could probably be taken almost
verbatim from the second midterm, though.